# docs/swagger.yaml
basePath: /api
definitions:
  models.AddressResponse:
    properties:
      suggestions:
        items:
          $ref: '#/definitions/models.AddressSuggestion'
        type: array
    type: object
  models.AddressSuggestion:
    properties:
      value:
        type: string
    type: object
  models.CreateFieldRequest:
    properties:
      address:
        type: string
      city:
        type: string
      description:
        type: string
      display:
        type: boolean
      dressing:
        type: boolean
      for_disabled:
        type: boolean
      info:
        type: string
      location:
        type: string
      logo:
        type: string
      media:
        type: string
      name:
        type: string
      parking:
        type: boolean
      places:
        type: integer
      slug:
        type: string
      square:
        type: integer
      toilet:
        type: boolean
    required:
    - address
    - city
    - name
    type: object
...
  models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/models.ValidationErrorResponse'
        type: array
      message:
        type: string
      statusCode:
        type: integer
    type: object
  models.FieldView:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      description:
        type: string
      display:
        type: boolean
      dressing:
        type: boolean
      for_disabled:
        type: boolean
      id:
        type: integer
      info:
        type: string
      location:
        type: string
      logo:
        allOf:
        - $ref: '#/definitions/models.Media'
        description: Логотип
      media:
        description: Медиа
        items:
          $ref: '#/definitions/models.Media'
        type: array
      name:
        type: string
      parking:
        type: boolean
      places:
        type: integer
      slug:
        type: string
      square:
        type: integer
      status:
        type: integer
      toilet:
        type: boolean
    type: object
...
  models.Media:
    properties:
      created_at:
        type: string
      ext:
        type: string
      id:
        type: integer
      name:
        type: string
      path:
        type: string
      size:
        type: integer
    type: object
...
  models.UpdateFieldRequest:
    properties:
      address:
        type: string
      city:
        type: string
      description:
        type: string
      display:
        type: boolean
      dressing:
        type: boolean
      for_disabled:
        type: boolean
      info:
        type: string
      location:
        type: string
      logo:
        type: string
      media:
        type: string
      name:
        type: string
      parking:
        type: boolean
      places:
        type: integer
      slug:
        type: string
      square:
        type: integer
      toilet:
        type: boolean
    required:
    - address
    - city
    - name
    type: object
...
  models.ValidationErrorResponse:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Приложения бронирования спортивных площадок.
  title: Спортивный город
  version: "1.0"
paths:
  /api/address/suggest:
    post:
      consumes:
      - application/json
      description: Получить подсказку по части адреса
      parameters:
      - description: Часть адреса для поиска
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AddressResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "415":
          description: Unsupported Media Type
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить подсказку по части адреса
      tags:
      - Адреса
...
  /api/fields:
    get:
      consumes:
      - application/json
      description: Получить список всех площадок
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FieldView'
            type: array
        "400":
          description: Bad Request
          schema:
            type: Bad
        "500":
          description: Internal Server Error
          schema:
            type: Internal
      summary: Получить все площадки
      tags:
      - Площадки
    post:
      description: Создать новую спортивную площадку с предоставленными данными
      parameters:
      - description: Данные для создания новой площадки
        in: body
        name: createField
        required: true
        schema:
          $ref: '#/definitions/models.CreateFieldRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FieldView'
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Создать новую площадку
      tags:
      - Площадки
  /api/fields/{id}:
    delete:
      description: Удалить спортивную площадку по её идентификатору
      parameters:
      - description: ID площадки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Площадка удалена
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Удалить площадку по ID
      tags:
      - Площадки
    put:
      description: Обновить существующую спортивную площадку с предоставленными данными
      parameters:
      - description: Данные для обновления площадки
        in: body
        name: updateField
        required: true
        schema:
          $ref: '#/definitions/models.UpdateFieldRequest'
      - description: ID площадки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FieldView'
        "404":
          description: Not Found
          schema:
            type: Not
        "422":
          description: Unprocessable Entity
          schema:
            type: Unprocessable
      summary: Обновить существующую площадку
      tags:
      - Площадки
  /api/fields/{slug}:
    get:
      description: Получить информацию о площадке по её slug
      parameters:
      - description: Slug площадки
        in: path
        name: slug
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FieldView'
        "400":
          description: Bad Request
          schema:
            type: Bad
        "404":
          description: Not Found
          schema:
            type: Not
      summary: Получить площадку по slug
      tags:
      - Площадки
...
swagger: "2.0"